move existing constraint to new package
